+++ netparams/snapshot.go
@@ -137,12 +137,9 @@ func (s *Store) LoadState(ctx context.Context, pl *types.Payload) ([]types.State
 	if !ok {
 		return nil, types.ErrInconsistentNamespaceKeys // it's the only possible key/namespace combo here
 	}
-	s.mu.Lock()
-	defer s.mu.Unlock()
+
 	for _, kv := range np.NetParams.Params {
-		if err := s.reloadKey(ctx, kv.Key, kv.Value); err != nil {
-			return nil, err
-		}
+		s.Update(ctx, kv.Key, kv.Value)
 	}
 
 	// Now they have been loaded, dispatch the changes so that the other engines pick them up
@@ -153,25 +150,4 @@ func (s *Store) LoadState(ctx context.Context, pl *types.Payload) ([]types.State
 	}
 
 	return nil, nil
-}
-
-// does the same as the regular store.Update call, but doesn't send an event.
-func (s *Store) reloadKey(ctx context.Context, key, value string) error {
-	svalue, ok := s.store[key]
-	if !ok {
-		// perhaps we should create one here?
-		// s.store[key] = value{}
-		return ErrUnknownKey
-	}
-
-	if err := svalue.Update(value); err != nil {
-		return fmt.Errorf("unable to update %s: %w", key, err)
-	}
-
-	// update was successful we want to notify watchers
-	s.paramUpdates[key] = struct{}{}
-	// set up the snapshot state to contain new data
-	s.state.update(key, value)
-
-	return nil
 }