+++ validators/topology_checkpoint_test.go
@@ -6,10 +6,10 @@ import (
 	"testing"
 
 	commandspb "code.vegaprotocol.io/protos/vega/commands/v1"
+	"code.vegaprotocol.io/vega/validators"
 	"github.com/stretchr/testify/assert"
 	abcitypes "github.com/tendermint/tendermint/abci/types"
 	types1 "github.com/tendermint/tendermint/proto/tendermint/types"
-	tmtypes "github.com/tendermint/tendermint/types"
 )
 
 func addNodes(top *testTop, number int) {
@@ -27,7 +27,7 @@ func addNodes(top *testTop, number int) {
 			ChainPubKey:     tmPubKeys[0],
 			VegaPubKey:      fmt.Sprintf("vega-key-%d", i),
 			EthereumAddress: fmt.Sprintf("eth-address-%d", i),
-		})
+		}, validators.ValidatorStatusTendermint)
 	}
 }
 
@@ -121,7 +121,7 @@ func testTopologyCheckpointUsesRelativeBlockHeight(t *testing.T) {
 	var newNetworkBlockHeight uint64 = 100
 
 	// set current block height to newNetworkBlockHeight
-	newTop.BeginBlock(ctx, abcitypes.RequestBeginBlock{Header: types1.Header{Height: int64(newNetworkBlockHeight)}}, []*tmtypes.Validator{})
+	newTop.BeginBlock(ctx, abcitypes.RequestBeginBlock{Header: types1.Header{Height: int64(newNetworkBlockHeight)}})
 
 	newTop.Load(ctx, ckp)